+++ client/internal/routemanager/manager_test.go
@@ -1,3 +1,5 @@
+//go:build privileged
+
 package routemanager
 
 import (
