+++ client/internal/routemanager/systemops/systemops_windows_test.go
@@ -1,3 +1,5 @@
+//go:build privileged
+
 package systemops
 
 import (
